Move spawn_handler
