+++ crates/kitsune_p2p/kitsune_p2p/src/gossip/sharded_gossip.rs
@@ -44,7 +44,7 @@ mod bandwidth;
 mod metrics;
 mod next_target;
 
-#[cfg(all(test, feature = "test_utils"))]
+#[cfg(test)]
 pub(crate) mod tests;
 
 /// max send buffer size (keep it under 16384 with a little room for overhead)
