+++ crates/kitsune_p2p/kitsune_p2p/src/gossip/sharded_gossip/tests.rs
@@ -13,8 +13,6 @@ mod handler_builder;
 mod test_local_sync;
 mod test_two_nodes;
 
-pub use common::spawn_handler;
-
 impl ShardedGossipLocal {
     pub fn test(
         gossip_type: GossipType,
